kaggle_amex_default/models/xgb_only_numeric/requirements.txt
wheel
joblib==1.2.0
numpy==1.23.4
pandas==1.5.0
pip==22.2.2
python-dateutil==2.8.2
pytz==2022.4
scikit-learn==1.1.2
scipy==1.9.2
setuptools==61.2.0
six==1.16.0
threadpoolctl==3.1.0
wheel==0.37.1
xgboost==1.6.2
sf-hamilton==1.10.0
PyYAML==5.4.1
typing-extensions==4.4.0
typing-inspect==0.8.0
pyarrow==10.0.1
